push some diagnostics
